[FE]_Modify can edit and delete comment

+++ frontend/src/app/components/comments/comments.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, Input, OnDestroy } from '@angular/core';
 import { CommentService } from '../../services/socket/comments.service';
 import { AuthService } from '../../services/auth.service';
 import { CommonModule } from '@angular/common';
@@ -11,9 +11,9 @@ import { AuthGuard } from '../../guards/auth.guard';
   standalone: true,
   imports: [CommonModule, FormsModule],
   templateUrl: './comments.component.html',
-  styleUrl: './comments.component.scss',
+  styleUrls: ['./comments.component.scss'],
 })
-export class CommentsComponent implements OnInit {
+export class CommentsComponent implements OnInit, OnDestroy {
   @Input() articleId: any;
   newComment: any = {
     idUser: '',
@@ -23,8 +23,10 @@ export class CommentsComponent implements OnInit {
   comments: any[] = [];
   rootComments: any[] = [];
   commentSubscription: Subscription | undefined;
-isDisplayMore: boolean = false;
+  isDisplayMore: boolean = false;
   router: any;
+  currentUserId: string = '';
+
   constructor(
     private commentService: CommentService,
     private readonly authService: AuthService,
@@ -32,14 +34,17 @@ isDisplayMore: boolean = false;
   ) {}
 
   ngOnInit(): void {
+    const accessTokenPayload = this.authService.getAccessTokenPayload();
+    this.currentUserId = accessTokenPayload ? accessTokenPayload.id : null;
+
     this.commentSubscription = this.commentService
       .onComments()
       .subscribe((comments: any[]) => {
-        this.rootComments = comments.filter(comment => comment.parentId === null);
-        console.log(comments);
-        
+        this.rootComments = comments.filter(
+          (comment) => comment.parentId === null
+        );
+
         this.comments = comments;
-        
       });
 
     this.commentService.requestComments(this.articleId);
@@ -49,41 +54,43 @@ isDisplayMore: boolean = false;
       if (newComment.parentId === null) {
         this.rootComments.push(newComment);
       }
-  
-      this.comments.push(newComment);
 
+      this.comments.push(newComment);
     });
 
-    this.commentService.onUpdateCommentSuccess().subscribe((updatedComment: any) => {
-      const index = this.comments.findIndex(
-        (c) => c._id === updatedComment._id
-      );
-      if (index !== -1) {
-        this.comments[index] = updatedComment;
-      }
-      
-    
-      if (updatedComment.parentId === null) {
-        const rootIndex = this.rootComments.findIndex(
+    this.commentService
+      .onUpdateCommentSuccess()
+      .subscribe((updatedComment: any) => {
+        const index = this.comments.findIndex(
           (c) => c._id === updatedComment._id
         );
-        if (rootIndex !== -1) {
-          this.rootComments[rootIndex] = updatedComment;
+        if (index !== -1) {
+          this.comments[index] = updatedComment;
         }
-      }
-    });
 
-    this.commentService.onDeleteCommentSuccess().subscribe((deletedComment: any) => {
-      this.comments = this.comments.filter(
-        (comment) => comment._id !== deletedComment._id
-      );
-      
-      if (deletedComment.parentId === null) {
-        this.rootComments = this.rootComments.filter(
+        if (updatedComment.parentId === null) {
+          const rootIndex = this.rootComments.findIndex(
+            (c) => c._id === updatedComment._id
+          );
+          if (rootIndex !== -1) {
+            this.rootComments[rootIndex] = updatedComment;
+          }
+        }
+      });
+
+    this.commentService
+      .onDeleteCommentSuccess()
+      .subscribe((deletedComment: any) => {
+        this.comments = this.comments.filter(
           (comment) => comment._id !== deletedComment._id
         );
-      }
-    });
+
+        if (deletedComment.parentId === null) {
+          this.rootComments = this.rootComments.filter(
+            (comment) => comment._id !== deletedComment._id
+          );
+        }
+      });
   }
 
   ngOnDestroy(): void {
@@ -92,10 +99,18 @@ isDisplayMore: boolean = false;
     }
   }
 
+  canEditComment(comment: any): boolean {
+    return comment.idUser._id === this.currentUserId;
+  }
+
+  canDeleteComment(comment: any): boolean {
+    return comment.idUser._id === this.currentUserId;
+  }
+
   sendNewComment(): void {
     if (!this.authGuard.canActivate()) {
-      this.router.navigate(['/login']); 
-      return; 
+      this.router.navigate(['/login']);
+      return;
     }
     const userId = this.authService.getAccessTokenPayload().id;
     this.newComment.idUser = userId;
@@ -105,6 +120,10 @@ isDisplayMore: boolean = false;
   }
 
   saveEdit(comment: any): void {
+    if (!this.canEditComment(comment)) {
+      console.error('User does not have permission to edit this comment.');
+      return;
+    }
     comment.isEditing = false;
     this.commentService.updateComment(comment).subscribe(
       (updatedComment: any) => {
@@ -135,6 +154,10 @@ isDisplayMore: boolean = false;
   }
 
   toggleEdit(comment: any) {
+    if (!this.canEditComment(comment)) {
+      console.error('User does not have permission to edit this comment.');
+      return;
+    }
     comment.originalContent = comment.content;
     comment.isEditing = true;
 
@@ -151,6 +174,11 @@ isDisplayMore: boolean = false;
   }
 
   deleteComment(commentId: string): void {
+    const comment = this.comments.find((c) => c._id === commentId);
+    if (comment && !this.canDeleteComment(comment)) {
+      console.error('User does not have permission to delete this comment.');
+      return;
+    }
     if (confirm('Are you sure you want to delete this comment?')) {
       this.commentService.deleteComment(commentId).subscribe(
         () => {},
@@ -171,23 +199,24 @@ isDisplayMore: boolean = false;
       idUser: userId,
       idArticle: this.articleId!,
       parentId: parentComment._id,
-      content: parentComment.newReplyContent, 
+      content: parentComment.newReplyContent,
     };
 
     this.commentService.sendComment(newReply);
-    parentComment.newReplyContent = ''; 
+    parentComment.newReplyContent = '';
     parentComment.isReplying = false;
   }
 
   cancelReply(parentComment: any) {
     parentComment.isReplying = false;
-    parentComment.newReplyContent = ''; 
+    parentComment.newReplyContent = '';
   }
+
   toggleDisplayMore(comment: any): void {
     comment.isDisplayMore = !comment.isDisplayMore;
   }
+
   getRepliesForComment(commentId: string): any {
-    return this.comments.filter(comment => comment.parentId === commentId);
+    return this.comments.filter((comment) => comment.parentId === commentId);
   }
-  
 }
